pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scanner", "scripts"]

[project]
name = "coding-agent-stats"
version = "0.1.0"
description = "Statistics on coding agent adoption across public GitHub repos"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
scan         = "scanner.__main__:run"
test-scan    = "scripts.test_scan:run"
serve        = "scripts.serve:run"
export-stats = "scripts.export_stats:run"
